# api/lex-openapi.yaml
...
openapi: 3.1.0
info:
  title: LEX — Lead Exchange Standard REST API
  version: 1.0.0
  description: |
    REST API for the LEX (Lead Exchange Standard) protocol.
    
    ## Overview
    LEX is a platform-agnostic specification for universal sales lead data exchange
    across industries including automotive, aviation, maritime, real estate,
    heavy equipment, and technology.
    This OpenAPI spec covers the **REST/Webhook transport layer** — the primary
    integration method for web-connected systems.
    
    ## Authentication
    LEX defines OAuth 2.0 Client Credentials (machine-to-machine) as the
    authentication pattern. Your platform implements the token endpoint.
    Include the Bearer token in the `Authorization` header on every request.
    
    ## Idempotency
    All POST endpoints are idempotent on `messageId`. Submitting the same
    `messageId` twice returns 200 OK with the original response — no duplicate
    processing occurs.
    
    ## Rate Limits
    - Standard tier: 1,000 requests/minute
    - Enterprise tier: 10,000 requests/minute
    - Batch endpoint: 100 messages/call
    
    ## Versioning
    The API version is in the URL path (`/v1/`). The LEX spec version is in
    each message's `header.version` field.
    
  contact:
    name: LEX Standard Working Group
    url: https://lexstandard.org/docs
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.lexstandard.org/v1
...
security:
  - oauth2: []
...
paths:
  /messages:
    post:
      tags: [messages]
      operationId: submitMessage
      summary: Submit a LEX message
      description: |
        Submit any LEX message type (LEAD, ASSET, ACKNOWLEDGMENT, SUBSCRIPTION, LEAD_CLOSURE).
        
        The server will:
        1. Validate the message envelope (header fields, message type)
        2. Apply schema validation for the payload
        3. Apply business rule validation
        4. Route to the receiver identified by `receiverId`
        5. Return the validation result synchronously
        
        **Idempotency:** Duplicate `messageId` submissions return 200 OK without
        reprocessing. This is the expected behavior for retry scenarios.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnyLexMessage'
            examples:
              lead:
                summary: LEAD message — customer expressing interest
                value:
                  lex:
                    header:
                      messageId: "MSG-2026-LEAD-001234"
                      messageType: LEAD
                      version: "1.0"
                      timestamp: "2026-03-25T09:00:00Z"
                      senderId: "PLATFORM-TOYOTA-US"
                      receiverId: "RETAILER-TOYOTA-FREMONT-001"
                      encryptionMethod: TLS1.3
                    payload:
                      lead:
                        leadId: "LEAD-2026-003321"
                        status: EXPRESSED_INTEREST
                        source: MANUFACTURER_WEBSITE
                        customer:
                          firstName: Marcus
                          lastName: Okafor
                          email: marcus.okafor@email.com
                          phone: "+14085559023"
                        desiredProduct:
                          productType: VEHICLE
                          manufacturers: [Toyota]
                          preferredModels: ["RAV4 Hybrid"]
                        metadata:
                          createdAt: "2026-03-25T09:00:00Z"
                          version: "1.0"
      responses:
        '200':
          description: Message processed (includes idempotent re-submissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageSubmitResponse'
        '400':
          description: Message validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailureResponse'
              example:
                messageId: "MSG-2026-LEAD-001234"
                valid: false
                errors:
                  - level: ERROR
                    field: "payload.lead.customer.email"
                    message: "Invalid email format"
                    code: INVALID_EMAIL_FORMAT
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /leads:
    post:
      tags: [leads]
      operationId: submitLead
      summary: Submit a LEAD message (type-specific shorthand)
      description: |
        Convenience endpoint for LEAD messages only. Equivalent to `POST /messages`
        with `messageType: LEAD` but with Lead-specific request/response shapes
        for easier client implementation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LexLeadRequest'
      responses:
        '200':
          description: Lead submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadSubmitResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional client-provided idempotency key. If provided, must match the
        `lex.header.messageId` in the request body. Used by clients to detect
        duplicate submissions at the HTTP layer before message parsing.
      schema:
        type: string
  schemas:
    AnyLexMessage:
      type: object
      required: [lex]
      properties:
        lex:
          type: object
          required: [header, payload]
          properties:
            header:
              type: object
              required: [messageId, messageType, version, timestamp, senderId, receiverId]
              properties:
                messageId:
                  type: string
                messageType:
                  type: string
                  enum: [LEAD, ASSET, ACKNOWLEDGMENT, SUBSCRIPTION, LEAD_CLOSURE]
                version:
                  type: string
                timestamp:
                  type: string
                  format: date-time
                senderId:
                  type: string
                receiverId:
                  type: string
                encryptionMethod:
                  type: string
                  enum: [NONE, TLS1.2, TLS1.3, END_TO_END]
                  default: TLS1.3
                correlationId:
                  type: string
                ttl:
                  type: object
                  properties:
                    maxAgeSeconds:
                      type: integer
                      default: 3600
                    expiresAt:
                      type: string
                      format: date-time
                    onExpiry:
                      type: string
                      enum: [DLQ, DISCARD, RETURN]
            payload:
              type: object
              description: Message-type-specific payload
    LexLeadRequest:
      type: object
      description: Standalone LEAD message request body
      required: [lex]
      properties:
        lex:
          type: object
          required: [header, payload]
          properties:
            header:
              $ref: '#/components/schemas/AnyLexMessage/properties/lex/properties/header'
            payload:
              type: object
              required: [lead]
              properties:
                lead:
                  type: object
                  description: See LEX_LEAD_SCHEMA.json for full schema
...
    MessageSubmitResponse:
      type: object
      required: [messageId, valid, processedAt]
      properties:
        messageId:
          type: string
        valid:
          type: boolean
        processedAt:
          type: string
          format: date-time
        processingTimeMs:
          type: integer
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        idempotent:
          type: boolean
          description: True if this was a duplicate submission — no reprocessing occurred
    LeadSubmitResponse:
      allOf:
        - $ref: '#/components/schemas/MessageSubmitResponse'
        - type: object
          properties:
            leadId:
              type: string
            routedTo:
              type: string
              description: LEX organization ID the message was routed to
    ValidationFailureResponse:
      type: object
      required: [messageId, valid, errors]
      properties:
        messageId:
          type: string
        valid:
          type: boolean
          example: false
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        processingTimeMs:
          type: integer
    ValidationIssue:
      type: object
      required: [level, field, message]
      properties:
        level:
          type: string
          enum: [CRITICAL, ERROR, WARNING]
        field:
          type: string
          description: JSON path to the problematic field (e.g., payload.lead.customer.email)
        message:
          type: string
        code:
          type: string
          description: Machine-readable error code
        suggestion:
          type: string
          description: Human-readable fix suggestion
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
  responses:
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: "Bearer token missing or expired"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit window resets
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: RATE_LIMITED
            message: "1000 requests/minute limit exceeded"
    ValidationFailed:
      description: Message validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationFailureResponse'
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: INTERNAL_ERROR
            message: "An unexpected error occurred. Request ID included for support."
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        OAuth 2.0 Client Credentials for machine-to-machine authentication.
        LEX defines this as the required auth pattern — your platform supplies
        the token endpoint. Replace the tokenUrl below with your auth server.
      flows:
        clientCredentials:
          tokenUrl: https://{your-auth-server}/oauth/token
          scopes:
            lex:leads:read: Read LEAD messages
            lex:leads:write: Submit and update LEAD messages
            lex:assets:read: Read ASSET messages
            lex:assets:write: Submit ASSET messages
            lex:closures:write: Submit LEAD_CLOSURE messages
            lex:subscriptions:manage: Create and manage subscriptions
            lex:conformance: Run conformance tests